requirements.txt
requests
bcrypt
anthropic
python-dotenv
openpyxl
